Revert clearing the input buffers in decoder
TC: testClearKeyPlaybackMpeg2ts is failing because Video playback is not
happening due to the commit: 984a09e10f0820b43cddfbe9810dc4508c8b0794 as
it is clearing the each incoming frame index's input buffer in the
decoder component.

Instead of clearing the input buffers initialize the m_uInputDelay in
decoder.

Tracked-On: OAM-106762
Signed-off-by: gollarx <ratnakumarix.golla@intel.com>